mark-flow-ui: Inline ElectionInfo component in StartPage
It was only being used here. It handled more cases than was
necesssary for various undefined props, so removed that handling as
well.
