+++ libs/mark-flow-ui/src/pages/start_page.tsx
@@ -7,16 +7,23 @@ import {
   AudioOnly,
   ReadOnLoad,
   PageNavigationButtonId,
+  Caption,
+  electionStrings,
+  H1,
+  NumberString,
+  P,
+  PrecinctSelectionName,
+  PrimaryElectionTitlePrefix,
+  Seal,
+  useScreenInfo,
 } from '@votingworks/ui';
-
-import { assert } from '@votingworks/basics';
-
+import { assert, assertDefined } from '@votingworks/basics';
 import {
   BallotStyleId,
   ElectionDefinition,
+  getBallotStyle,
   PrecinctId,
 } from '@votingworks/types';
-import { ElectionInfo } from '../components/election_info';
 import { ContestsWithMsEitherNeither } from '../utils/ms_either_neither_contests';
 import { VoterScreen } from '../components/voter_screen';
 
@@ -28,6 +35,15 @@ const wobbleKeyframes = keyframes`
   97% { transform: rotate(6deg); }
   98% { transform: rotate(-1deg); }
   99% { transform: rotate(2deg); }
+`;
+
+const ElectionInfo = styled.div`
+  display: flex;
+  gap: 0.5rem;
+
+  @media (orientation: portrait) {
+    flex-direction: column;
+  }
 `;
 
 const StartVotingButton = styled(Button)`
@@ -68,6 +84,63 @@ export function StartPage(props: StartPageProps): JSX.Element {
     'ballotStyleId is required to render StartPage'
   );
 
+  const { election } = electionDefinition;
+  const { county, seal } = election;
+  const screenInfo = useScreenInfo();
+
+  const electionInfo = (
+    <ElectionInfo>
+      <Seal
+        seal={seal}
+        maxWidth="7rem"
+        style={{
+          marginRight: screenInfo.isPortrait ? undefined : '1rem', // for horizontal layout
+          marginBottom: screenInfo.isPortrait ? '0.5rem' : undefined, // for vertical layout
+        }}
+      />
+      <div>
+        <H1>
+          <PrimaryElectionTitlePrefix
+            ballotStyleId={ballotStyleId}
+            election={election}
+          />
+          {electionStrings.electionTitle(election)}
+        </H1>
+        <P>{electionStrings.electionDate(election)}</P>
+        <P>
+          <Caption maxLines={4}>
+            {/* TODO(kofi): Use more language-agnostic delimiter (e.g. '|') or find way to translate commas. */}
+            <span>
+              <PrecinctSelectionName
+                electionPrecincts={election.precincts}
+                precinctSelection={singlePrecinctSelectionFor(precinctId)}
+              />
+              ,{' '}
+            </span>
+            {electionStrings.countyName(county)},{' '}
+            {electionStrings.stateName(election)}
+          </Caption>
+          <Caption>
+            {appStrings.labelBallotStyle()}{' '}
+            {electionStrings.ballotStyleId(
+              assertDefined(
+                getBallotStyle({
+                  ballotStyleId,
+                  election,
+                })
+              )
+            )}
+          </Caption>
+          <br />
+          <Caption>
+            {appStrings.labelNumBallotContests()}{' '}
+            <NumberString value={contests.length} />
+          </Caption>
+        </P>
+      </div>
+    </ElectionInfo>
+  );
+
   const startVotingButton = (
     <StartVotingButton
       variant="primary"
@@ -83,12 +156,7 @@ export function StartPage(props: StartPageProps): JSX.Element {
     <VoterScreen padded>
       <div style={{ margin: 'auto', padding: '0.5rem' }}>
         <ReadOnLoad>
-          <ElectionInfo
-            electionDefinition={electionDefinition}
-            ballotStyleId={ballotStyleId}
-            precinctSelection={singlePrecinctSelectionFor(precinctId)}
-            contestCount={contests.length}
-          />
+          {electionInfo}
           <AudioOnly>{introAudioText}</AudioOnly>
         </ReadOnLoad>
         {startVotingButton}
